refactored exports and added ability to switch visible data type
