+++ src/Form/Type/ReportSumType.php
@@ -1,0 +1,58 @@
+<?php
+
+/*
+ * This file is part of the Kimai time-tracking app.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace App\Form\Type;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\OptionsResolver\Options;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
+
+class ReportSumType extends AbstractType
+{
+    private $authorizationChecker;
+
+    public function __construct(AuthorizationCheckerInterface $authorizationChecker)
+    {
+        $this->authorizationChecker = $authorizationChecker;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults([
+            'required' => true,
+            'multiple' => false,
+            'expanded' => true,
+        ]);
+
+        $resolver->setDefault('choices', function (Options $options) {
+            $choices = ['stats.durationTotal' => 'duration'];
+
+            if ($this->authorizationChecker->isGranted('view_rate_other_timesheet')) {
+                $choices['stats.amountTotal'] = 'rate';
+                $choices['label.rate_internal'] = 'internalRate';
+            }
+
+            return $choices;
+        });
+
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getParent()
+    {
+        return ChoiceType::class;
+    }
+}
