+++ src/Reporting/AbstractUserList.php
@@ -11,11 +11,9 @@ namespace App\Reporting;
 
 abstract class AbstractUserList
 {
-    /**
-     * @var \DateTime
-     */
     private $date;
     private $decimal = false;
+    private $sumType = 'duration';
 
     public function getDate(): ?\DateTime
     {
@@ -35,5 +33,19 @@ abstract class AbstractUserList
     public function setDecimal(bool $decimal): void
     {
         $this->decimal = $decimal;
+    }
+
+    public function getSumType(): ?string
+    {
+        return $this->sumType;
+    }
+
+    public function setSumType(string $sumType): void
+    {
+        if (!\in_array($sumType, ['duration', 'rate', 'internalRate'])) {
+            throw new \InvalidArgumentException('Unknown sum type');
+        }
+
+        $this->sumType = $sumType;
     }
 }
