+++ src/Reporting/WeekByUserForm.php
@@ -9,6 +9,7 @@
 
 namespace App\Reporting;
 
+use App\Form\Type\ReportSumType;
 use App\Form\Type\UserType;
 use App\Form\Type\WeekPickerType;
 use Symfony\Component\Form\AbstractType;
@@ -41,6 +42,7 @@ class WeekByUserForm extends AbstractType
         if ($options['include_user']) {
             $builder->add('user', UserType::class, ['width' => false]);
         }
+        $builder->add('sumType', ReportSumType::class);
     }
 
     /**
