+++ tests/Controller/Reporting/ReportByUserControllerTest.php
@@ -30,19 +30,19 @@ class ReportByUserControllerTest extends ControllerBaseTest
 
     public function testWeekByUserIsSecure()
     {
-        $this->assertUrlIsSecured('/reporting/week_by_user');
+        $this->assertUrlIsSecured('/reporting/user/week');
     }
 
     public function testMonthByUserIsSecure()
     {
-        $this->assertUrlIsSecured('/reporting/month_by_user');
+        $this->assertUrlIsSecured('/reporting/user/month');
     }
 
     public function testUserWeekReport()
     {
         $client = $this->getClientForAuthenticatedUser(User::ROLE_SUPER_ADMIN);
         $this->importReportingFixture(User::ROLE_SUPER_ADMIN);
-        $this->assertAccessIsGranted($client, '/reporting/week_by_user?user=4&date=12999119191');
+        $this->assertAccessIsGranted($client, '/reporting/user/week?user=4&date=12999119191');
         self::assertStringContainsString('<div class="box-body user-week-reporting-box', $client->getResponse()->getContent());
         $option = $client->getCrawler()->filterXPath("//select[@id='user']/option[@selected]");
         self::assertEquals(4, $option->attr('value'));
@@ -52,7 +52,7 @@ class ReportByUserControllerTest extends ControllerBaseTest
     {
         $client = $this->getClientForAuthenticatedUser(User::ROLE_USER);
         $this->importReportingFixture(User::ROLE_USER);
-        $this->assertAccessIsGranted($client, '/reporting/month_by_user?user=4&date=12999119191');
+        $this->assertAccessIsGranted($client, '/reporting/user/month?user=4&date=12999119191');
         self::assertStringContainsString('<div class="box-body user-month-reporting-box', $client->getResponse()->getContent());
         $select = $client->getCrawler()->filterXPath("//select[@id='user']");
         self::assertEquals(0, $select->count());
